requirements.txt
flake8==7.0.0
mccabe==0.7.0
mysqlclient==2.2.4
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
setuptools==69.5.1
shortuuid==1.0.13
